requirements.txt
numpy == 1.21.0
scikit-learn==0.22.1
Pillow==7.0.0